// models/productSchema.ts
export interface WooProduct {
  id?: string;
  type?: string;
  sku: string;
  name: string;
  published?: number;
  isFeatured?: boolean;
  visibility?: string;
  shortDescription?: string;
  description?: string;
  dateOnSaleFrom?: string;
  dateOnSaleTo?: string;
  taxStatus?: string;
  taxClass?: string;
  inStock?: boolean;
  stock?: number;
  lowStockAmount?: number;
  backordersAllowed?: boolean;
  soldIndividually?: boolean;
  weight?: number;
  dimensions?: { length: number; width: number; height: number };
  allowReviews?: boolean;
  purchaseNote?: string;
  regularPrice?: number;
  salePrice?: number;
  categories?: string[];
  tags?: string[];
  shippingClass?: string;
  images?: string[];
  downloadLimit?: number;
  downloadExpiryDays?: number;
  parent?: string;
  groupedProducts?: string[];
  upsells?: string[];
  crossSells?: string[];
  externalUrl?: string;
  buttonText?: string;
  position?: number;
  attributes?: { name: string; value: string; visible: boolean; global: boolean }[];
  metaData?: { key: string; value: string }[];
}